+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Not finding a private entry when opening in the web viewer
 - Incorrect color of the "Close" web viewer button in dark theme
 - Missing focus indication in file selection dialog
+- Unwanted focus on the "Back" toolbar button in the media viewer
 
 ## [1.15.0] - 2023-07-18